% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test5-package.R
\docType{package}
\name{test5-package}
\alias{test5}
\alias{test5-package}
\title{test5: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Ramnath Vaidyanathan \email{ramnath.vaidya@gmail.com}

}
\keyword{internal}
